Weight unit: store kg (Double) throughout, remove all g/kg conversions

+++ src/main/java/com/zilai/zilaibuy/dto/parcel/ParcelDto.java
@@ -16,7 +16,7 @@ public record ParcelDto(
         BigDecimal declaredValue,
         String status,
         String processingOption,
-        Integer weight,
+        Double weight,
         String outboundTrackingNo,
         String notes,
         Long linkedOrderId,
